Migration alteração grupos pessoas e processo seletivo

+++ database/migrations/2018_08_27_000018_create_group_person_table.php
@@ -24,7 +24,7 @@ class CreateGroupPersonTable extends Migration
                 $table->foreign('status_id')->references('id')->on('statuses');
                 $table->integer('status_aso_id')->unsigned()->nullable();
                 $table->foreign('status_aso_id')->references('id')->on('statuses');
-                $table->integer('note')->nullable();
+                $table->decimal('note', 8, 1)->nullable();
                 $table->text('description')->nullable();
                 $table->timestamps();
             });
